statistics integration
- count core is now different component
- team list now shows only teams with 3+ games, displays games quantity
- add team is now a component, not a route
- new icons
