+++ src/components/add-team-name/add-team-name.component.tsx
@@ -11,14 +11,14 @@ import { selectCurrentUser } from '~/store/user/user.selector';
 export const AddTeamName = () => {
   const dispatch = useDispatch();
   const currentUser = useSelector(selectCurrentUser);
-  const [teamName, setTeamName] = useState('');
+  const [teamName, setTeamName] = useState(currentUser?.teamName);
   const { t } = useTranslation('translation', { keyPrefix: 'auth' });
 
   if (!currentUser) return <Navigate to='/' />;
 
   const handleSubmit = async (event: FormEvent<HTMLFormElement>) => {
     event.preventDefault();
-    dispatch(addTeamName(teamName));
+    if (teamName) dispatch(addTeamName(teamName));
   };
 
   const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
@@ -27,28 +27,23 @@ export const AddTeamName = () => {
   };
 
   return (
-    <div className='sign-in'>
-      <div className='sign-in__wrapper'>
-        <h3>{t('addTeam')}</h3>
-        <form onSubmit={handleSubmit}>
-          <FormInput
-            required
-            label={t('teamName')}
-            name='team-name'
-            type='text'
-            onChange={handleChange}
-            value={teamName}
-          />
-          {teamName.toLocaleLowerCase() === 'admin' || teamName.toLocaleLowerCase() === '...' ? (
-            <h2>{t('wrongTeam')}</h2>
-          ) : (
-            <Button type='submit' buttonType={BUTTON_CLASSES.apply}>
-              {t('add')}
-            </Button>
-          )}
-        </form>
-        {currentUser?.teamName && <Navigate to='/' />}
-      </div>
-    </div>
+    <form onSubmit={handleSubmit}>
+      <FormInput
+        required
+        label={t('teamName')}
+        name='team-name'
+        type='text'
+        onChange={handleChange}
+        value={teamName}
+      />
+      {teamName &&
+      (teamName.toLocaleLowerCase() === 'admin' || teamName.toLocaleLowerCase() === '...') ? (
+        <span>{t('wrongTeam')}</span>
+      ) : (
+        <Button type='submit' buttonType={BUTTON_CLASSES.apply}>
+          {t('add')}
+        </Button>
+      )}
+    </form>
   );
 };
