+++ src/routes/navigation/navigation.styles.scss
@@ -4,10 +4,7 @@
   align-items: center;
   display: grid;
   grid-template-columns: 1fr 1fr;
-  grid-auto-rows: 35px;
-  @include bp(tab) {
-    grid-auto-rows: 45px;
-  }
+  grid-auto-rows: 45px;
   @include bp(laptop) {
     grid-auto-rows: 60px;
   }
@@ -36,10 +33,13 @@
     display: flex;
     justify-content: end;
     align-items: center;
-    gap: 20px;
+    gap: 10px;
     padding: 0 10px;
+    @include bp(tab) {
+      gap: 20px;
+    }
     & svg {
-      height: 20px;
+      height: 28px;
     }
   }
   & button {
